Avoid re-rendering the Modal component
I noticed that the Modal component was duplicated everytime the application would cause a trigger of componentDidUpdate(), resulting (in my case) in a huge amount of modals coming out of nowhere.
After gathering some information about Preact, I found several issues on the main repo where the author explains that the default behaviour of React.render() in Preact is NOT "replace" but append.
To make the Modal replace itself during re-renders, we pass a third parameter to React.render in ModalPortal.js